requirements.txt
google-adk
google-genai
pandas
numpy
python-dotenv